Refactored to API and Views routes; added approve, deny and delete buttons to approval page
