+++ public/js/global.js
@@ -16,7 +16,7 @@ $(document).ready(function() {
 	$('#logout').click(function() {
 		$.ajax({
 			method: "POST",
-			url: "/logout",
+			url: "/api/logout",
 			success: function() {
 				window.location.href = '/login';
 			}
