+++ views/error.jade
@@ -2,5 +2,6 @@ extends layout
 
 block content
   h1= message
-  h2= error.status
+  h2 An error has occurred.
+  h2 Code: #{(error.status ? error.status : "None")}
   pre #{error.stack}
