+++ views/login.jade
@@ -6,7 +6,7 @@ block styles
 
 block content
   .login-container
-    form.form-signin(action='/login', method='post')
+    form.form-signin(action='/api/login', method='post')
       h2.form-signin-heading
         p Login
