+++ views/register.jade
@@ -5,7 +5,7 @@ block styles
     link(href='/css/register.css', rel='stylesheet')
 block content
     .login-container
-      form.form-signin(action='/register', method='post')
+      form.form-signin(action='/api/register', method='post')
         h2.form-signin-heading
           p Register
